+++ src/cli/invoke.ts
@@ -12,7 +12,7 @@ import mri from 'mri';
 
 import type {Args} from '../task/task.js';
 import {invokeTask} from '../task/invokeTask.js';
-import {nodeFilesystem} from '../fs/node.js';
+import {fs as nodeFs} from '../fs/node.js';
 
 /*
 
@@ -32,7 +32,7 @@ const main = async () => {
 	} = argv;
 	const args = {_, ...namedArgs};
 
-	await invokeTask(nodeFilesystem, taskName, args);
+	await invokeTask(nodeFs, taskName, args);
 };
 
 main(); // see `attachProcessErrorHandlers` above for why we don't catch here