+++ src/fs/filesystem.ts
@@ -1,13 +1,13 @@
-import type {PathLike, Stats, CopyOptions, MoveOptions} from 'fs-extra';
-
-import type {PathStats} from './pathData.js';
+import {stripEnd} from '../utils/string.js';
+import type {Flavored} from '../utils/types.js';
+import type {Encoding} from './encoding.js';
+import type {PathData, PathStats} from './pathData.js';
 import type {PathFilter} from './pathFilter.js';
 
-export type {Stats} from 'fs-extra';
-
+// API is modeled after `fs-extra`: https://github.com/jprichardson/node-fs-extra/
 export interface Filesystem {
 	stat: FsStat;
-	pathExists: FsPathExists;
+	pathExists: stringExists;
 	readFile: FsReadFile;
 	readJson: FsReadJson;
 	outputFile: FsOutputFile;
@@ -21,29 +21,32 @@ export interface Filesystem {
 }
 
 export interface FsStat {
-	(path: PathLike): Promise<Stats>;
+	(path: string): Promise<PathStats>;
 }
-export interface FsPathExists {
+export interface stringExists {
 	(path: string): Promise<boolean>;
 }
 export interface FsReadFile {
-	(file: PathLike | number): Promise<Buffer>;
-	(file: PathLike | number, encoding: string): Promise<string>;
+	// TODO ran into problems using overrides, couldn't get subclass to typecheck:
+	(path: string): Promise<Buffer>;
+	(path: string, encoding: 'utf8'): Promise<string>;
+	(path: string, encoding: null): Promise<Buffer>;
+	(path: string, encoding?: Encoding): Promise<Buffer | string>;
 }
 export interface FsReadJson {
-	(file: string): Promise<any>;
+	(path: string): Promise<any>;
 }
 export interface FsOutputFile {
-	(file: string, data: any, encoding?: string): Promise<void>;
+	(path: string, data: any, encoding?: Encoding): Promise<void>;
 }
 export interface FsRemove {
 	(path: string): Promise<void>;
 }
 export interface FsMove {
-	(src: string, dest: string, options?: MoveOptions): Promise<void>; // TODO which options?
+	(src: string, dest: string, options?: FsMoveOptions): Promise<void>; // TODO which options?
 }
 export interface FsCopy {
-	(src: string, dest: string, options?: CopyOptions): Promise<void>; // TODO which options? all? breaks conventionn with above
+	(src: string, dest: string, options?: FsCopyOptions): Promise<void>; // TODO which options? all? breaks conventionn with above
 }
 export interface FsReadDir {
 	(path: string): Promise<string[]>;
@@ -61,4 +64,87 @@ export interface FsFindFiles {
 		// pass `null` to speed things up at the risk of infrequent misorderings (at least on Linux)
 		sort?: ((a: [any, any], b: [any, any]) => number) | null,
 	): Promise<Map<string, PathStats>>;
+}
+
+// TODO maybe swap these two names?
+export abstract class Fs implements Filesystem {
+	abstract stat: FsStat;
+	abstract pathExists: stringExists;
+	abstract readFile: FsReadFile;
+	abstract readJson: FsReadJson;
+	abstract outputFile: FsOutputFile;
+	abstract remove: FsRemove;
+	abstract move: FsMove;
+	abstract copy: FsCopy;
+	abstract readDir: FsReadDir;
+	abstract emptyDir: FsEmptyDir;
+	abstract ensureDir: FsEnsureDir;
+	abstract findFiles: FsFindFiles;
+}
+
+// TODO remove some of these, but which?
+export interface FsCopyOptions {
+	dereference?: boolean;
+	overwrite?: boolean;
+	preserveTimestamps?: boolean;
+	errorOnExist?: boolean;
+	filter?: FsCopyFilterSync | FsCopyFilterAsync;
+	recursive?: boolean;
+}
+export interface FsMoveOptions {
+	overwrite?: boolean;
+	limit?: number;
+}
+export type FsCopyFilterSync = (src: string, dest: string) => boolean;
+export type FsCopyFilterAsync = (src: string, dest: string) => Promise<boolean>;
+
+export class FsStats implements PathStats {
+	constructor(private readonly _isDirectory: boolean) {}
+	isDirectory() {
+		return this._isDirectory;
+	}
+}
+
+export type FsId = Flavored<string, 'FsId'>;
+
+// TODO improve and move
+// name? `toNormalPath`? `normalize`? `normalizePath`?
+export const toFsId = (path: string): FsId => stripEnd(path, '/');
+
+// TODO extract - how do they intersect with Filer types? and smaller interfaces like `PathData`?
+export type FsNode = TextFileNode | BinaryFileNode | DirectoryNode;
+
+// TODO should we use `BaseFilerFile` here?
+// this mirrors a lot of that, except this models directories as well,
+// which we wanted to upstream ..
+// so this is a good entrypoint for backporting dirs into the Filer
+
+// TODO names? should they have `Fs` prefix? extract?
+
+export interface BaseNode {
+	readonly id: FsId;
+	readonly isDirectory: boolean;
+	readonly encoding: Encoding;
+	readonly contents: string | Buffer | null;
+	// readonly contentsBuffer: Buffer | null;
+	readonly stats: PathStats;
+	readonly path: PathData; // TODO name? `PathInfo`? `PathMeta`? `Path`?
+}
+export interface BaseFileNode extends BaseNode {
+	readonly isDirectory: false;
+	readonly contents: string | Buffer;
+}
+export interface TextFileNode extends BaseFileNode {
+	readonly encoding: 'utf8';
+	readonly contents: string;
+}
+export interface BinaryFileNode extends BaseFileNode {
+	readonly encoding: null;
+	readonly contents: Buffer;
+	// readonly contentsBuffer: Buffer;
+}
+export interface DirectoryNode extends BaseNode {
+	readonly isDirectory: true;
+	readonly contents: null;
+	// readonly contentsBuffer: null;
 }