+++ src/fs/inputPath.test.ts
@@ -12,7 +12,7 @@ import {
 import type {PathStats} from './pathData.js';
 import {groPaths, replaceRootDir, createPaths, paths} from '../paths.js';
 import type {Obj} from '../index.js';
-import {nodeFilesystem} from './node.js';
+import {fs} from './node.js';
 
 /* test_resolveRawInputPath */
 const test_resolveRawInputPath = suite('resolveRawInputPath');
@@ -126,7 +126,7 @@ const test_loadSourcePathDataByInputPath = suite('loadSourcePathDataByInputPath'
 test_loadSourcePathDataByInputPath('loads source path data and handles missing paths', async () => {
 	const result = await loadSourcePathDataByInputPath(
 		{
-			...nodeFilesystem,
+			...fs,
 			pathExists: async (path) => path !== 'fake/test3.bar.ts' && !path.startsWith('fake/missing'),
 			stat: async (path) =>
 				({